[#166] refactor: 홈페이지 폴더구조 변경
