+++ src/app/_features/home/components/HomeMainContent.tsx
@@ -1,0 +1,15 @@
+import EventAndCalendarPanel from './EventAndCalendarPanel';
+import HomeTopImage from './HomeTopImage';
+import BasicAndPatchPanel from './notice/BasicAndPatchPanel';
+
+const HomeMainContent = () => {
+  return (
+    <div className="dark flex flex-1 flex-col gap-2 p-2">
+      <HomeTopImage />
+      <BasicAndPatchPanel />
+      <EventAndCalendarPanel />
+    </div>
+  );
+};
+
+export default HomeMainContent;
