+++ src/app/_features/home/components/HomeTopImage.tsx
@@ -1,0 +1,13 @@
+const HomeTopImage = () => {
+  return (
+    <div
+      className="relative h-60 w-full rounded-md bg-cover"
+      style={{
+        backgroundImage: 'url(/art.jpg)',
+        backgroundPosition: 'center 12%',
+      }}
+    />
+  );
+};
+
+export default HomeTopImage;
