+++ src/app/_features/home/index.ts
@@ -1,0 +1,1 @@
+export { default as HomeMainContent } from './components/HomeMainContent';
